Fixed minor bug in MoveToAttack
Since position 0,0 is commonly the center of the map, and often times valid, it's technically possible for an actor to stand on that exact position, in which case MoveToAttack would consider the target to have moved every time. By adding an explicitly invalid position, we can more easily tell whether a position was set or not, which may come in handy in other situations as well.