Added server conf client script
Exposes the world conf to the client, so we can easily adjust client behaviors and UIs based on our configuration. I don't expect this to ever become a security concern, because there are no sensitive information in the world conf, but alternatively we could limit the list of options to ones we know to be safe.