Refactored buff handlers
OnStart is getting replaced by OnActivate, a method that better describes what is happening and that comes with a parameter that tells us whether the skill was started or overbuffed. This should make it easier for newcomers to understand how the buff handlers operate and allows us to more easily adjust the behavior based on how the buff was activated.