Lowered max size for conf client script table size
It's not clear how, but it was apparently possible to trigger the max length check, even though the table is currently only ~1900 bytes in Melia and it should be difficult to add over a hundred bytes with a single option value. But since it happened, we're gonna reduce the max to 1900 and see what happens. Technically we should be checking if the table will be above the max _before_ adding to it, but that's not of vital importance and this fix should do it for now.