Added global main group chat
Adds a chat room that is created by default that anybody can join. Using the >main command displays an invite tag for the group chat that can be clicked to easily join it. Afterwards it's a normal group chat that you can also leave again, filter, customize, etc. It's currently unknown how scalable the group chats are and whether the client would be able to handle dozens and hundreds of players in one chat, but for the time being this seems like a neat way of handling a main chat, as compared to handling it via chat or system messages.