Minor change in Throwing Spear pad explosion logic
Removed an unnecessary method and replaced the max target retrival, to use the pad for it instead. It has all the logic for it already, which will come in handy if we have to make any changes to it, such as adding an option to override all max targets based on a potential "DisablePadMaxTarget" check. We shouldn't use the SDR option for it, as that's a different mechanic.