Refactored server events
Replaced them to use the new Event class from Yggdrasil. Another awesome change for *all* the WinMergers out there, as this requires changing every single event triggering and subscribing =) The reasoning for this change is that the default events aren't great for our use case, as they require dedicated methods to raise them, creating quite a bit of boilerplate that needs to be maintained. Additionally, this gives us a bit more control over the events, more options to optimize them in the future, and more explicit (un)subscriptions compared to the +=/-= format.