Added override look up for scriptable functions
Basic support for getting the original function in cases where scripts override existing scriptable functions. For example, a custom script might override a skill calculation function to modify the result passed to the server. With access to the original function, the custom script can retrieve the original return value and apply changes to it, such as multiplying it, without needing to recreate the function or even know what the original did.