Added basic registration web page
Simple one-page registration that utilizes the web API. It's a bit of a proof of concept, as you might not actually want to use the API in a production server, just as you wouldn't want to use the prefix-based account creation, since they're difficult to control and rate-limit. But it might be a good temp solution for a small server.