Fixed cookie handling in PHP
We were setting headers by index when sending them back to the client, which resulted in some cookies being omitted if there was more than one. This commit changes it so cookie headers are added to the existing list. It's possible that we'll need this for more headers down the line, or that we might need to change how the headers are transferred entirely, but that's a problem for future-exec. Another simple solution might be to just turn full control over to PHP and ignore whatever EmbedIO set.