Updated unclutter script
Added removal of stat reset button in character info. We don't currently support it and its design is heavily targeting players of officials. If users end up wanting this functionality via button, we could add an option for it. Though it seems easier to solve it via NPC or command.